pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel", "pip>=22.1"]
build-backend = "setuptools.build_meta"

[project]
name = "security-ghost"
version = "0.1.0"
description = "A security tool for anonymous networking"
authors = [{name = "Aaron Wilmoth"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "requests",
    "pysocks",
    "python-wireguard",
    "python-decouple",
]

[project.scripts]
security-ghost = "security_ghost.skeleton:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["security_ghost*", "lib*"]
namespaces = false

[tool.setuptools.package-data]
security_ghost = ["mac-vendor.txt"]